refactor : LottoMoney 변수 타입을 long으로 변경

+++ src/main/java/lotto/domain/result/LottoRank.java
@@ -48,7 +48,7 @@ public enum LottoRank {
 		return matchCount.getMatchCount();
 	}
 
-	public int getWinningLottoMoney() {
+	public long getWinningLottoMoney() {
 		return winningLottoMoney.getLottoMoney();
 	}
